+++ src/core.rs
@@ -1,6 +1,5 @@
 pub mod bus;
 pub mod db;
-pub mod error;
 pub mod message;
 pub mod si;
 pub mod value;