+++ src/services/buienradar.rs
@@ -49,7 +49,7 @@ impl Buienradar {
             .station_measurements
             .iter()
             .find(|measurement| measurement.station_id == self.station_id)
-            .ok_or_else(|| format!("station {} is not found", self.station_id))?;
+            .ok_or_else(|| anyhow!("station {} is not found", self.station_id))?;
         let sensor_prefix = format!("{}::{}", service_id, self.station_id);
         if let Some(temperature) = measurement.temperature {
             Message::new(format!("{}::temperature", sensor_prefix))
@@ -185,7 +185,7 @@ struct BuienradarStationMeasurement {
 }
 
 /// Implements [custom date/time format](https://serde.rs/custom-date-format.html) with Amsterdam timezone.
-fn deserialize_datetime<'de, D: Deserializer<'de>>(deserializer: D) -> Result<DateTime<Local>, D::Error> {
+fn deserialize_datetime<'de, D: Deserializer<'de>>(deserializer: D) -> StdResult<DateTime<Local>, D::Error> {
     Ok(Amsterdam
         .datetime_from_str(&String::deserialize(deserializer)?, "%Y-%m-%dT%H:%M:%S")
         .map_err(de::Error::custom)?